requirements.txt
accelerate==1.1.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
albucore==0.0.20
albumentations==1.4.21
all-clip==1.2.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
async-timeout==5.0.1
attrs==24.2.0
autofaiss==2.17.0
blinker==1.9.0
blis==1.0.1
braceexpand==0.1.7
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
clip-anytorch==2.6.0
clip-retrieval==2.44.0
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.10
dataclasses==0.6
datasets==3.1.0
diffusers==0.31.0
dill==0.3.8
docker-pycreds==0.4.0
embedding-reader==1.7.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
ExifRead-nocycle==3.0.1
faiss-cpu==1.9.0.post1
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
fire==0.5.0
Flask==3.1.0
Flask-Cors==4.0.2
Flask-RESTful==0.3.10
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
gitdb==4.0.11
GitPython==3.1.43
gradio==4.44.1
gradio_client==1.3.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
img2dataset==1.45.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
memory-profiler==0.61.0
mpmath==1.3.0
multidict==6.1.0
multilingual-clip==1.0.10
multiprocess==0.70.16
murmurhash==1.0.11
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
open_clip_torch==2.29.0
opencv-python-headless==4.10.0.84
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==10.4.0
preshed==3.0.9
prometheus_client==0.21.0
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
pyarrow==14.0.2
pydantic==2.10.1
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.8.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.11.4
semantic-version==2.10.0
sentence-transformers==2.7.0
sentry-sdk==2.19.0
setproctitle==1.3.4
shellingham==1.5.4
simsimd==6.1.1
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.41.3
stringzilla==3.10.10
sympy==1.13.1
termcolor==2.5.0
thinc==8.3.2
threadpoolctl==3.5.0
timm==1.0.11
tokenizers==0.20.3
tomlkit==0.12.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.46.3
triton==3.1.0
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
wandb==0.16.6
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.100
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.0
zipp==3.21.0
